+++ week4/Testcode.py
@@ -1,11 +1,11 @@
-def factorial_recursive(n):
-    if n == 0 or n == 1:  # กรณีฐาน (base case)
-        return 1
-    else:
-        return n * factorial_recursive(n - 1)  # เรียกฟังก์ชันซ้ำโดยลด n ลง 1
+# def factorial_recursive(n):
+#     if n == 0 or n == 1:  # กรณีฐาน (base case)
+#         return 1
+#     else:
+#         return n * factorial_recursive(n - 1)  # เรียกฟังก์ชันซ้ำโดยลด n ลง 1
 
 # ตัวอย่างการเรียกใช้
-print(factorial_recursive(5))
+# print(factorial_recursive(5))
 
 def pongkarm(n):
     if n == 0:
@@ -14,4 +14,3 @@ def pongkarm(n):
         return n + pongkarm(n - 1)
 
 print(pongkarm(9))
-
