+++ week4/teach_sort/EX_Quick_sort.py
