+++ week5/Homework/Maze_V2.py
@@ -21,38 +21,38 @@ class maze:
         # self.end = pos(2, 6)
         # self.end = pos(5, 1)
         # self.ply = pos(2, 6)
-        # self.maze = [
-        #             ["X", "X", "X", "X", "X", "X", "X"],
-        #             ["X", " ", " ", " ", "X", " ", "X"],
-        #             ["X", " ", "X", " ", "X", " ", "X"],
-        #             ["X", " ", "X", " ", "X", " ", "X"],
-        #             ["X", " ", "X", " ", " ", " ", "X"],
-        #             ["X", " ", "X", " ", "X", " ", " "],
-        #             ["X", " ", "X", " ", "X", " ", "X"],
-        #             ["X", "X", "X", " ", "X", " ", "X"],
-        #             ["X", " ", " ", " ", "X", " ", "X"],
-        #             ["X", " ", "X", " ", "X", " ", "X"],
-        #             ["X", " ", "X", "X", "X", "X", "X"],
-        #             ]
-        # self.ply = pos(10, 1)
-        # self.end = pos(5, 6)
-        # self.end = pos(10, 1)
-        # self.ply = pos(5, 6)
         self.maze = [
-                    ["X", "X", "X", "X", "X", "X", "X", " ", "X"],
-                    ["X", " ", " ", " ", "X", " ", "X", " ", "X"],
-                    ["X", " ", "X", " ", " ", " ", "X", " ", "X"],
-                    ["X", " ", "X", " ", "X", " ", "X", " ", "X"],
-                    ["X", " ", " ", " ", "X", " ", "X", " ", "X"],
-                    ["X", " ", "X", " ", "X", " ", "X", " ", "X"],
-                    ["X", " ", "X", " ", "X", " ", "X", " ", "X"],
-                    ["X", " ", "X", " ", "X", " ", "X", " ", "X"],
-                    ["X", " ", " ", " ", "X", " ", " ", " ", "X"],
-                    ["X", " ", "X", " ", " ", " ", "X", " ", "X"],
-                    ["X", " ", "X", "X", "X", "X", "X", "X", "X"],
+                    ["X", "X", "X", "X", "X", "X", "X"],
+                    ["X", " ", " ", " ", "X", " ", "X"],
+                    ["X", " ", "X", " ", "X", " ", "X"],
+                    ["X", " ", "X", " ", "X", " ", "X"],
+                    ["X", " ", "X", " ", " ", " ", "X"],
+                    ["X", " ", "X", " ", "X", " ", " "],
+                    ["X", " ", "X", " ", "X", " ", "X"],
+                    ["X", "X", "X", " ", "X", " ", "X"],
+                    ["X", " ", " ", " ", "X", " ", "X"],
+                    ["X", " ", "X", " ", "X", " ", "X"],
+                    ["X", " ", "X", "X", "X", "X", "X"],
                     ]
         self.ply = pos(10, 1)
-        self.end = pos(0, 7)
+        self.end = pos(5, 6)
+        # self.end = pos(10, 1)
+        # self.ply = pos(5, 6)
+        # self.maze = [
+        #             ["X", "X", "X", "X", "X", "X", "X", " ", "X"],
+        #             ["X", " ", " ", " ", "X", " ", "X", " ", "X"],
+        #             ["X", " ", "X", " ", " ", " ", "X", " ", "X"],
+        #             ["X", " ", "X", " ", "X", " ", "X", " ", "X"],
+        #             ["X", " ", " ", " ", "X", " ", "X", " ", "X"],
+        #             ["X", " ", "X", " ", "X", " ", "X", " ", "X"],
+        #             ["X", " ", "X", " ", "X", " ", "X", " ", "X"],
+        #             ["X", " ", "X", " ", "X", " ", "X", " ", "X"],
+        #             ["X", " ", " ", " ", "X", " ", " ", " ", "X"],
+        #             ["X", " ", "X", " ", " ", " ", "X", " ", "X"],
+        #             ["X", " ", "X", "X", "X", "X", "X", "X", "X"],
+        #             ]
+        # self.ply = pos(10, 1)
+        # self.end = pos(0, 7)
         self.maze[self.ply.y][self.ply.x] = "P"
         self.maze[self.end.y][self.end.x] = "E"
     #เช็คว่าอยู่ในบอร์ดอยู่หรือเปล่า
@@ -130,7 +130,7 @@ class maze:
                 time.sleep(0.25)
         return True
 
-    def lookway(self):
+    def lookway(self,stack):
         #up
         if self.maze[self.ply.y-1][self.ply.x] != "X" and self.maze[self.ply.y-1][self.ply.x] != 'o' and self.maze[self.ply.y-1][self.ply.x] != 'i':
             way = 1
@@ -145,69 +145,69 @@ class maze:
             way = 4    
         else:
             way = 5                 
-        return way
-    def go_to_way(self, way):
+        return stack, way
+    def go_to_way(self,stack, way):
         if way == 1:
             pk.move_up()
-            # stack.push('p')
+            stack.push('p')
         elif way == 2:
             pk.move_left()
-            # stack.push('p')
+            stack.push('p')
         elif way == 3:
             pk.move_right()
-            # stack.push('p')
+            stack.push('p')
         elif way == 4:
             pk.move_down()
-            # stack.push('p')
+            stack.push('p')
         elif way == 5:
+            #return way
+            # stack.pop()
             if self.maze[self.ply.y-1][self.ply.x] == "i" or self.maze[self.ply.y-1][self.ply.x] == 'E':
                 next_move = pos(self.ply.y-1, self.ply.x)
                 self.maze[self.ply.y][self.ply.x] = "o"
                 self.maze[next_move.y][next_move.x] = "P"
-                # self.ply = next_move
+                self.ply = next_move
             elif self.maze[pk.ply.y][self.ply.x-1] == "i" or self.maze[self.ply.y][self.ply.x-1] == 'E':
                 next_move = pos(self.ply.y, self.ply.x-1)
                 self.maze[self.ply.y][self.ply.x] = "o"
                 self.maze[next_move.y][next_move.x] = "P"
-                # self.ply = next_move
+                self.ply = next_move
             elif self.maze[self.ply.y][self.ply.x+1] == "i" or self.maze[self.ply.y][self.ply.x+1] == 'E' :
                 next_move = pos(self.ply.y, self.ply.x+1)
                 self.maze[self.ply.y][self.ply.x] = "o"
                 self.maze[next_move.y][next_move.x] = "P"
-                # self.ply = next_move
+                self.ply = next_move
             elif self.maze[self.ply.y+1][self.ply.x] == "i" or self.maze[self.ply.y+1][self.ply.x] == 'E':
                 next_move = pos(self.ply.y+1, self.ply.x)
                 self.maze[self.ply.y][self.ply.x] = "o"
                 self.maze[next_move.y][next_move.x] = "P"
-                # self.ply = next_move   
+                self.ply = next_move   
            
-        return #stack
+        return stack
 class pos: 
     def __init__(self, y=None, x=None):
         self.y = y
         self.x = x
 
-#no stack
-#ออกด้วยวิธีการทำสัญลักษณ์บนแผนที่
+
 if __name__ == '__main__':
 
     pk = maze()
-    # stack = Stack()
+    stack = Stack()
     print('Press Enter to start')
     while True:
         if keyboard.is_pressed("enter"):
             pk.print()
             break
 
-    # stack.push('p')
+    stack.push('p')
     while True:
         if keyboard.is_pressed("q"):
             print("Quit Program")
             break
-        way = pk.lookway()
-        pk.go_to_way(way)
+        stack, way = pk.lookway(stack)
+        stack = pk.go_to_way(stack, way)
         pk.print()
-        # time.sleep(0.1)
+        time.sleep(0.1)
 
-#bug look way
 #No chat GPT !!!!!!!!!!!
